restaurant-review-system/facial_emotion/requirements_efficientnet.txt
torch>=2.0.0
torchvision>=0.15.0
timm
Pillow
numpy
matplotlib
opencv-python
